chore: add user's origin cluster name to audit log
While investigating #12324, I realized that our audit logs do not
include any metadata associated with the user's provenence. This means
in root-leaf clusters, it's impossible to figure out if a user was a
remote user or a local (leaf) user by looking into the audit logs.

This PR adds a new field `user_cluster_name` which holds the Teleport
cluster name where the user is authenticated against.

In leaf clusters, this will be the root cluster name. For root clusters
or single clusters, the field will hold the local cluster name.