Update github.com/sigstore/rekor to v1.4.1
This update allows us to update go-chi/chi to v5.2.2 which resolves
GHSA-vrw8-fxc6-2r93.

The update also forces us to bump testify to 1.11.0 which brings
with it a change in behavior of require.Eventually. In prior versions,
the assertion function was initially only called _after_ the first tick.
In 1.11.0, the assertion function is called _immediately_ and then
on all subsequent ticks. This change in behavior has brought to light
a number of existing tests making use of require.Eventually with
buggy assertion functions. Their bugs were masked by the fact that
the intial tick was always enough time for the assertion to be
successful. The remediations in this change may not be all inclusive
as not all require.Eventually uses were audited, but all tests which
failed in one way or another are addresssed.