Add Terraform modules for deploying Access Graph with Teleport. Support local deployment requiring `terrafrom`, `aws` and `mkcert` CLIs. Modules generate access-graph and teleport yaml configs as well as `docker-compose.yaml` file to orchestrate running the Access Graph and Teleport services.
Create self-signed, internal certificates for Access Graph and the
Teleport proxy certificate with `mkcert` so that it is registered with the
system trust store and we can test on localhost and via Browser.

Add docs and sample root module.

In follow up commits we will add EC2 deployment, Identity Activity
Center support and entries in the official Teleport docs.