[v18] Enable Docusaurus tagging in docs frontmatter (#58818)
* Enable Docusaurus tagging in docs frontmatter

Backports #58660

Replace the `labels` frontmatter field with one called `tags`.
Docusaurus supports tag-based navigation out of the box with the `tags`
field. The original plan was to implement a separate system that allows
for filtering by multiple tags - which Docusaurus does not - but it
makes sense to enable the default approach as a first iteration.

* Tag Identity Governance pages: use case and cloud

Tag pages in the Identity Governance section of the docs with the
appropriate use case and cloud provider tags. Cloud provider and use
case are common categories that matter for Teleport users, so we can
implement these tags across the docs to allow readers to find useful
content.