Allow control of service account.
The current behavior is to generate YAML with no `serviceAccountName`
nor `automountServiceAccountToken` set. The Kubernetes default is to use
the "default" service account, and mount the token. The principle of least
privilege suggests that the pod in this chart should not get a token,
because it does not interact with the Kubernetes API.