Tag guides by cloud provider and use case (#58867)
Tag pages in the following sections of the docs with the appropriate use
case and cloud provider tags.

- Connect your client
- Identity Security
- Installation

Cloud provider and use case are common categories that matter for
Teleport users, so we can implement these tags across the docs to allow
readers to find useful content.

Note that not all pages in these sections warrant a use case or cloud
provider tag.