Remove tags from the docs home page
The docs home page is a special case, the first page a user sees when
they visit the docs site (i.e., without navigating to a specific page
from a search engine). As a result, it does not make sense to display
this page in lists of pages with a particular tag, since in that case,
it would not be the first page a user sees.