Tag three Enroll Resources docs sections
Tag pages in the following sections of the docs with the appropriate use
case and cloud provider tags:

- Enroll Applications
- Auto-Discovery
- Enroll Servers

Cloud provider and use case are common categories that matter for
Teleport users, so we can implement these tags across the docs to allow
readers to find useful content.

This change also fixes missing tags.

Note that not all pages in these sections warrant a use case or cloud
provider tag.