[AWSIC] Address flakey group import tests (#7201)
* [AWSIC] Address flakey group import tests

It the default OIDC authentication used by the AWS Identity Center
integration tests blocks test execution for ~1.5s while it times
out. This eats half of the 3s timeouts on various test conditions
on integration startup, and the remaining 1.5s is not always enough
time for the conditions to be met.

This patch updates the affected AWSIC tests to use the ambient system
credentials, which do not have the same timeout, giving more tme for
the test conditions to be met.

This patch also:
 - Adds tooling to make it easier to tests to find fixtures when run
   under docker for local flakiness testing,
 - Removes an assertion from `TestAWSGroupImportCreatesAccessLists`
   that is immaterial to the test at hand.

Addresses: gravitational/teleport.e#7198

* Remove fixture stuff

* Use ambient AWS creds by default

Source-commit: gravitational/teleport.e@1e6d0944edf931f302546de09c09968b76480e58
Source-pull-requests: gravitational/teleport.e#7201