Marshal iterator into event instead of map[string]any (#59030)
This commit fixes a data loss bug causing the DynamoDB cursor EventIndex
field to be sightly changed due to conversion issues. As this field is
used to index events, this could lead to paginated queries not returning
the right events, either returning events from before or after the
requirested page. In the worst case, this could cause a livelock as the
query continuisly processes the same events.

The data loss issue is caused by improper JSON unmarshalling of large
integers. This happened because of this reasons:
- JSON is fundamentally flawed as it offers a single number type "binary64"
  for all numbers, whether they are integers or float. Go's
  encoding/json library uses field types to detect if the number should
  be stored in an int64 or a float64.
- [The AWS SDK v2 migration PR](https://github.com/gravitational/teleport/pull/44363)
  changed the cursor JSON unmarshalling logic and unmarshalled the
  cursor into `map[string]any`. This caused every integer field of
  `event` to round-trip through float64.
- [The Emit event fallback PR](https://github.com/gravitational/teleport/pull/40854)
  changed the EventIndex value from a small incremental integer to a
  large unix nanosecond timestamp in case of conflict. The large value
  was no longer safe for storage in a float64.

The combination of those 3 factors caused the cursor EventIndex to get
corrupted and caused unexpected event query index offsets. When preseted
with a non-existing document, DynamoDB still hashes it and starts the
query from its supposed location in the index. This is why this issue
has not been detected for so long. Its consequences were:
- duplicated events returned on 2 consecutive pages (this case was
  handled properly by the event forwarder as it keeps track of the last
  processed event)
- livelock if the number of duplicated events exceed the page size
- non-forwarded events if the index offset was in the future