Remove ParseAccessList key and add title sort index
We don't need to parse the access list nextkey if we know how to create
it when we send to the client. now, we just use whatever key is sent to
us, and we create the key as needed based on sort index (which we have
at time of request). also, the web UI needs a title index (we sort by
title in the web UI, not name) so I added that. added toLower to sort in
a logical way (similar to unified resources) and encoding to protect
against special characters in the title (an should preserve
lexicographical sorting)