Stop loading webassets in Proxy API tests
The Proxy API tests relied on the presence of compiled
webassets in a specific relative directory. This prevented
the tests from being executed outside of the lib/web directory.

None of the tests however actually require webassets, and the
Proxy is capable of serving operating without webassets, allowing
the removal of webassets from the tests entirely.