Add tags to four Enroll Resources docs sections (#58979)
Tag pages in the following sections of the Zero Trust Access docs with
the appropriate use case and cloud provider tags:

- Agents
- Enroll Desktops
- Enroll Kubernetes Clusters
- MCP

Cloud provider and use case are common categories that matter for
Teleport users, so we can implement these tags across the docs to allow
readers to find useful content.

This change also corrects pages with missing or inaccurate tags.