[v18] deflake `TestResizeTerminal` (#59163)
* deflake `TestResizeTerminal`

On `TestResizeTerminal` test flow, a session is created by a user foo.
Once the session is created, the user foo sends a rezize event to set
the terminal size to `100x100`.

Once the session tracker is available, another user "bar" joins the
session and sends two resize events - one for joining and another
because it also sets the terminal size to 100x100.

To avoid resize loops, Teleport prevents the resize sender to receive
the resize event it generated. Thus, this test validates if the user
"bar" never receives the two resize events it triggers when joining the
session.

The issue arrises from the fact that the code didn't ensure that the
initial resize event triggered by user foo was properly handled before
bar joins. This means that bar was receiving a resize event that foo
triggered before he joined the session but the server only processed it
afterwards as you can see in the log bellow:

```
    apiserver_test.go:1970:
        	Error Trace:	/__w/teleport/teleport/lib/web/apiserver_test.go:1970
        	Error:      	terminal 2 should not have received a resize event: %v
        	Test:       	TestResizeTerminal
        	Messages:   	Version:"1" Type:"a" Payload:"{\"cluster_name\":\"localhost\",\"code\":\"T2002I\",\"ei\":3,\"event\":\"resize\",\"login\":\"root\",\"namespace\":\"default\",\"private_key_policy\":\"none\",\"server_addr\":\"127.0.0.1:44619\",\"server_hostname\":\"node\",\"server_id\":\"node\",\"server_version\":\"18.2.1\",\"sid\":\"b14a7676-4e43-48a0-b260-7ff5379d1790\",\"size\":\"100:100\",\"time\":\"2025-09-15T13:52:51.217Z\",\"uid\":\"01e9ca35-27d9-428d-ad20-5c9b43267b0f\",\"user\":\"foo\",\"user_cluster_name\":\"localhost\",\"user_kind\":1,\"user_roles\":[\"user:foo\"]}"
```

This PR deflakes the test by ensuring we catch the initial resize event
before bar joins the session.

Fixes #13607

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* Update apiserver_test.go

Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: rosstimothy <39066650+rosstimothy@users.noreply.github.com>