[v17] Adds services for pruning roles when requesting remote resources. (#59063)
Currently, when a user creates an Acess Request for a resource in a remote cluster
Teleport lacks the appropriate information to decide which roles to offer and falls
back to suggesting all available `search_as` roles for the reequesting user.

This patch provides the back-end services to allow Teleport to prune the roles for
remote-resource requests in the same way as requests for local resources.

Also note that this change only includes includes the back-end changes that will be
executed on a leaf cluster. The front-end, root-cluster changes are coming in a
subsequent PR.

Addresses: #56315
Backports: #57227
Changelog: Added services to correctly choose Access Request roles in remote clusters.