Server Auto Discover: handle suffix'd installations (#59094)
Teleport can be installed using `teleport-update`.
When installing teleport with a suffix (ie `teleport-update -i`) it will
install teleport in non-default locations, as well as an expectation to
use non-default locations for data directory and configuration and a
different systemd unit name.

In this scenario, running `teleport install autodiscover` should use
those non-default values.

This PR changes that command to consider those, and not the default ones
when the `TELEPORT_INSTALL_SUFFIX` env var is set.