deflake `TestMFAAuthenticateChallenge_IsMFARequiredApp` test (#59182)
* deflake `TestMFAAuthenticateChallenge_IsMFARequiredApp` test

`TestMFAAuthenticateChallenge_IsMFARequiredApp` relies on root-leaf
cluster to test MFA requirements for different applications.

The issue happens when the leaf cluster is able to upsert itself into
root auth server but it's not yet fully connected to root proxy and so
when the proxy tried to dial back to retrieve the registered app
servers, the request failed.

This PR fixes that by waiting for the root cluster proxy to be able to
dial to leaf auth server to retrive the leaf's application servers.

Fixes #59148

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>

* Update service.go

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

* Update app_aws_test.go

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

---------

Signed-off-by: Tiago Silva <tiago.silva@goteleport.com>
Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>