Add tags to two Zero Trust Access docs sections (#59136)
Tag pages in the following sections of the Zero Trust Access docs with
the appropriate use case and cloud provider tags:

- Compliance Frameworks
- Management

Cloud provider and use case are common categories that matter for
Teleport users, so we can implement these tags across the docs to allow
readers to find useful content.

Note that not all pages in these sections warrant a use case or cloud
provider tag.

This change also corrects pages with missing or inaccurate tags.