discovery: Add AWS EKS audit log fetching for Access Graph
Add a watcher to start fetchers for all access graph EKS clusters that
are configured to have Kubernetes apiserver audit logs fetched and send
them to access graph.

This is currently incomplete in that it does not have any dynamic
behaviour; DiscoveryConfig resources are not handled, nor do we update
the fetchers if the discovered clusters change. That will need a
reconciliation process to start/stop fetchers as the config and clusters
change.

More thought needs to be given to clusters coming and going. We
currently rely on access graph to maintain the resumption cursor, and
that only gets sent when the connection to access graph is performed. If
that connection is maintained and a cluster goes away and then comes
back, we might want to maintain an resumption state ourselves.

This commit requires go.mod be updated with:

    go get github.com/aws/aws-sdk-go-v2/service/cloudwatchlogs@latest

It is left out of this commit for now as it makes rebasing/merging
master easier.