Simplify Kubernetes Operator reference index page
The current list of guides is long and difficult to navigate. Instead,
include a description of the section, recommend reading the Managing
Resources guides for typical patterns, and indicate that there is one
guide per resource version.

Another possibility would be to make the list of reference docs easier
to navigate by formatting it as a table of resource names and
descriptions. However, we currently do not include descriptions of
resources when generating CRDs from resource proto messages.