Clarify configs in self-hosted Event Handler guides (#56937)
* Edit self-hosted Event Handler guides

Closes #38184

Prevent users from using example Event Handler configurations without
modifying the placeholder values. Remove placeholder values from example
Event Handler configs and add them to the comments accompanying the
modified fields.

Exceptions are if the accompanying guide does not explain the
placeholder, e.g., `serviceType` and the `log` section. In these cases,
we expect users to copy the placeholders and use them as they are.

Skip purely hosted Event Handler guides, which don't include
instructions for configuring the Event Handler.

Note that in the MS Teams guide, the TOML example includes clear
placeholders to indicate format, and this change leaves these in.

Also note that some guides, like the PagerDuty guide, include an example
of the final config files after providing lists of instructions. This
change does not touch the example configs in this case.

* Respond to kopiczko feedback

- Fill in missing examples.
- Consistently place comments above configuration fields.
- Use Cloud domain in the email Helm example.