Autogenerate the Upgrading sidebar section (#59252)
Backports #57993

Since this section has an increasing, regularly changing set of pages,
we can autogenerate it to make sure that no pages in the section go
without a sidebar.

Assign a `sidebar_position` of 1 to the Upgrading Compatibility
Overview, since it is a conceptual guide that we expect users to be
familiar with.