MWI: Add `tbot copy-binaries` subcommand (#59110) (#59404)
* MWI: Add `copy-binaries` subcommand

This adds a trivial `tbot copy-binaries <destination>` subcommand to
copy tbot (and optionally `fdpass-teleport`) to a given destination.

This command is useful for (among other possible use cases) copying
binaries from a Kubernetes initContainer so tbot or fdpass can be
executed from that environment. This is currently impossible to do
using our distroless images as they do not have a usable shell.

Additionally, this includes `fdpass-teleport` in our tbot-distroless
image, which currently does not include it.

* Add missing cli/copy_binaries.go

* Fix imports