docs: revamp Getting Started Guide alternative in one page
The docs revamp initially started out as a 4-page guide (#57931). This PR is an alternative condensed into one page.

Instead of the Getting Started page being a guide to setting up Enterprise cloud and resources, this PR updates the getting started page to help users choose an edition, deploy a cluster, connect infrastructure, setting up asccess, and configuring audit logging in a concise all in one page with appropriate links out to more in depth material if needed.