event-hander: skip desktop recording events by default (#59651)
The desktop.recording audit event is the analogous to the session
"print" event - it is emitted to session recordings but not to
the audit log, and contains the recorded session data.

The event handler was configured to skip SSH/Kube recording events
by default, but not desktop events. Desktop recording events are
even less useful for forwarding since they are mostly binary data.