Add a short intro paragraph to the changelog (#59745)
This resolves the expectation in the `remark-lint-page-structure` linter
that each page begin with at least one introductory paragraph. The
changes in #59542 attempted to ignore the linter, but since the `remark`
plugin for including partials preserves line numbers from the original
partial file in included AST nodes, the comment directive that ignores
the linter takes place after the linter violation. Placing the comment
before the violation causes the frontmatter to render incorrectly.

Merging gravitational/docs-website#393 removes the extraneous H1 heading
below the first paragraph.