Update bintest to 3.3.0 (#59815)
Some tests, which use bintest, were failing because of a potential lock.

This test hasn't failed in other major branches (v17, v18 and master)
which are using bintest 3.3.0.

Updating the version used in v16 should bring more stability to tests.

Given that this package is only used in tests, should be safe to update
even on a old release branch.
No side effects in production code are expected.