ci: Update create-github-app-token bloat.yaml workflow
Update create-github-app-token in bloat.yaml to use the latest v2. The
comment there says not to go past 1.0.5 as that is the lastest that is
supported on Centos7, but the buildbox used as the container for the
jobs is not a Centos7 container.

Bump it up and see if it works.

Also change underscores to hyphens in the input parameters as the
underscore version was removed in v2.