ci: Unpin old actions in bloat.yaml workflow
Update create-github-app-token in bloat.yaml to use the latest v2. The
comment there says not to go past 1.0.5 as that is the lastest that is
supported on Centos7, but the buildbox used as the container for the
jobs is not a Centos7 container.

Also change underscores to hyphens in the input parameters as the
underscore version was removed in v2.

Bump the version of actions/checkout to the latest too as that was
pinned to an old version for the same reason.