Disable package manager cache in setup-node action
The upgrade to v5 of actions/setup-node has set `package-manager-cache`
to `true` by default. This ends up requiring pnpm already be installed
for some reason. Some people pre-install pnpm with an install-pnpm
action but we already have a post-install action to run `corepack enable
pnpm`. We do not want another specialised action to just install some
node thing waiting for a supply chain attack. So just disabled the cache
as it was before and get on with things.