Clean up some lib/utils code (#59790)
* Clean up some lib/utils code

Rename CloserConn to WaitConn. This utility had the ability to tie
multiple io.Closers to the lifecycle of a net.Conn, but we never
used it for that purpose, and instead use it only as a way to wait
for a connection to be closed.

Remove the unnecessary utils.Stater interface. The only usage of
this interface already happened behind a type assertion, so there
was no need for an interface.

Un-export TrackingReader and TrackingWriter since they are
implementation details of TrackingConn.

* Move SyncBuffer to integration test package