Add a way to announce which sevrices should be expected (#59667)
I was looking into tying the auth readiness with its cache health (so we
don't end up in a state with no ready cache across all auths during a
rollout) and I saw that we are currently reporting ready as soon as one
of the sevrice heartbeats. We don't keep track of which services should
be heartbeating/reporting ready.

This PR introduces a new `process.ExpectService(component)` function to
declare early that we are starting a service and that the process should
not be ready without it.