Revert connectrpc.com/connect Go module to v1.18.1 (#59940)
Revert the Go module `connectrpc.com/connect` to v1.18.1 from v1.19.0 as
the latter version introduces a bug that causes build failures on 32-bit
platforms. This breaks our build for arm and 386.

The bug has been fixed upstream, but the PR is still open/unmerged after
a week, so it is not clear what the priority of 32-bit builds is for
connect. There would also need to be a new release of the Go module,
bumping the version. So for now, we just revert the version bump that
dependabot did.