EC2 Auto Discovery: retry matchers
After migrating to aws sdk v2, we started to fail fast when the AWS
credentials were not valid for a given AWS Matcher.

This means that the service exits when there's an AWS error.
This is not the case in v17 branch, where we still use aws sdk v1.

This new behavior also means that if we have 1000 discovery matchers,
each with its own assume role, and there's only one that has an invalid
assume role, it will exit instead of processing the other 999.