Deflake TestTrackingSession
When running the test at high count (1000) and
the CPU is heavily loaded the resource release
can be delayed enough to race with TempDir cleanup
leading to unlinkat errors in tests.

This change adds a short delay to give time for
remaing file locks to be released after closing
the db.