[cache] generic range implementation refactor (#60038)
This change modifies the generic range for cache
implmeneations to chunk the read of items such that
no single stream keeps the read lock for the duration
of the read.

This also fixes the issue in listRange which previously
would not respect the end token when falling back to upstream.

When falling back on a get all method, this implementation now
also clips the start and end to respect the method paramethers.