[v17] Clarify self-hosted Enterprise starter guidance (#59444)
* Clarify self-hosted Enterprise starter guidance

Closes #8547

Make it more explicit in the self-hosted Teleport Enterprise index page
that self-hosted deployents start with outreach to the Teleport team,
and that the HA guide and multi-region blueprint are outlines to help a
team plan a deployment that a customer designs in conversation with the
Teleport team.

Otherwise, the implication is that the guides listed on the self-hosted
Teleport Enterprise index page are sufficient for deploying a
self-hosted cluster. There has been a longstanding request for Oracle
Cloud guidance, and rather than expect these users to be self
sufficient, explain that the recommended approach is to reach out to
Teleport and model a deployment based on our architectural outlines.

Some new text in the self-hosted Enterprise index page is based
on #57931.

* Edit the self-hosted Teleport Enterprise intro

Respond to internal feedback by **pschisa**. The multi-region blueprint
is recommended only for a select group of organizations with specific
needs. For the general introduction, link only to the single-region HA
guide.