Improve IAM requirement visibility in EC2 guide
Closes #60077

Edit the EC2 instance auto-discovery guide so the list of IAM
permissions required for discoverable EC2 instances is in the
Prerequisites section. There is no reason for this discussion to be in
the same section of the page as the permissions for the Teleport
Discovery Service itself. Since attaching an IAM policy to discoverable
instances takes place outside the scope of this guide, it makes sense to
add this information to the Prerequisites.