Stop using the sqlite backend in tests
Swaps the lite backend for the memory backend in all tests that
can allow it. This could lead to reducing test times and is a
workaround for https://github.com/gravitational/teleport/issues/60127
- which is being addressed seperately.

The _only_ tests which have a hard requirement to use the sqlite backend
those which create a teleport process via service.NewTeleport. These
processes validate the the backend type is _not_ memory as that's not
a valid backend for a real Teleport process.