Allow `machineid.AutoUpdateVersionReporter` to shut down correctly (#60219)
* Allow `machineid.AutoUpdateVersionReporter` to shut down correctly

Fixes an issue where, because we were starting the leader election routine
inside `auth.NewServer`, nothing was waiting for it to shut down correctly,
and we were failing to release the semaphore.

There wasn't a huge user-facing impact (as the semaphore lease would eventually
expire) but it leads to a confusing error message in the logs and delays things
unnecessarily.

* Use `Clock.Since` instead of manually subtracting time

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>

---------

Co-authored-by: Zac Bergquist <zac.bergquist@goteleport.com>