fix: Ensure a SecReportsService is always registered (#7410)
OSS will not register the service if it detects that access monitoring is enabled. However, there were two cases in enterprise where a tenant was licensed for access monitoring, but the configuration did not support access monitoring that would return early without registering a service. This leads the cache to be unhealthy and in an infinite loop of fetching resources.

```

{"timestamp":"2025-10-13T14:26:39Z","level":"warning","caller":"auth/plugin.go:559","message":"Access Monitoring Enabled but Athena backend is not configured","component":"auth.enterprise"}
{"timestamp":"2025-10-14T12:58:07Z","level":"warning","caller":"cache/cache.go:1042","message":"Re-init the cache on error","component":"proxy:1:cache","target":"proxy","error":"failed to fetch resource: \"security_report_state\"\n\tunknown service teleport.secreports.v1.SecReportsService"}
```

Source-commit: gravitational/teleport.e@f36bdf3aac3321f1afa144dd7cf40fcd0d98d767
Source-pull-requests: gravitational/teleport.e#7410