Reduce time to report health changes from a Kubernetes cluster
The time to report a health change from a Kubernetes cluster to the auth server and UI is reduced to the `HealthCheckConfig.Spec.Interval` of `30s`.

Adding a health comparison to function `compareKubernetesServers` effects the call chain `(*kubeServerHeartbeatV2).Poll()` -> `GetServerInfo`, and reduces the communication time. `GetServerInfo` is called every `5s`, and notices changes updated by the `healthcheck` package every `30s`. Prior to this change, health was reported approximately every `6m` during the `(*kubeServerHeartbeatV2).Announce()` function.

Changes:
- Compare health status in function `compareKubernetesServers`

Part of #58413