AWS IAM Roles Anywhere: ignore sync profile when listing profiles (#60250)
The AWS IAM Roles Anywhere integration syncs Profiles as Teleport AWS
Apps.

In order to use the API, Teleport uses a specific Profile which is
configured as the Profile Syncer's Profile.
This Profile is only meant to be used by Teleport for syncing profiles,
and should not be available as an end-user Profile/Teleport AWS App.

During the Sync set up, we show the user all the available profiles.
This listing operation, wrongly, returns the Profile used for the
Profile Syncer.

We are showing a Profile that will not be available later on.

This PR changes the listing to only show the profiles that will be
synced, removing the Profile used in the Profile Syncer.