Add structure to the IP allowlist guide
As we add caveats and observations about the IP allowlist to
`docs/pages/ips.mdx`, including each new note in an Admonition makes the
guide more difficult to read. This change adds H2s to the guide so we
can place related notes next to one another without giving the user
colored boxes to skip over.