Improve IaC docs index page link visibility (#59461)
The index page of the Infrastructure as Code section of the docs was
originally intended to explain dynamic resources, so links to the guides
contained within the section are neither up to date nor particularly
visible. This change improves the visibility of links to the sections
within the IaC section, and focuses the IaC index page more on its
current purpose.

Edit the IaC index page to:
- Explain dynamic resources in terms of IaC tools.
- Remove information that is already included elsewhere, and not
  relevant to using IaC tools to manage Teleport.
- Update reference links. Add links to Operator and Terraform provider
  resource references.
- Move key guide links to the introductory section of this page.
- Remove the example manifests, since these aren't necessary once a user
  chooses to get started with a particular IaC tool. The examples are
  too minimal to be useful, but long enough to take up substantial space
  within the guide. This helps shorten the guide and focus on key
  information.