Use Integration audience when generating OIDC JWT
Prior to this change, the OIDC JWT generator uses a hardcoded `discover.teleport`
audience value. This change updates the token generator to use the configured
audience value from the supplied Integration resource, falling back to the
existing default if either no Integration is specified, or no audience is
specified in the Integration.