Update Audit page to use server side search/pagination
This PR will enable the audit page to be fully server side searchable
and paginated. I removed the sort by "type" because we don't have that
sort of searching in the backend, but you can still search/filter by it,
so you don't lose much functionality. I purposefully did not use
infinite query here to keep it as close to the current implementation as
possible, but i implemented the server side search with our new
react-query useInfiniteQuery hook so when we do update it again to
infinite scroll, it'll be easy. added some missing tests as well.