[api] Add ListSAMLConnectors (#60465)
This commits deperecates GetSAMLConnectors, which should be replaced by a paginated counterpart:
ListSAMLConnectors.

For client call sites two methods are added:
`ListSAMLConnectorsWithOptions`
`RangeSAMLConnectorsWithOptions`

These should replace `GetSAMLConnectors` and
`GetSAMLConnectorsWithValidationOptions`

Note that as tracked by gravitational/teleport/issues/60456;
the backend should not make external URL calls when unmarshalling
items. It is therefore expected that these method signatures may change in the
future when this is addressed, hence only the `*WithOptions` variants
are added to allow easier migration in the future. This is also the
reason for preserving the current method parameters as they are.

Towards: gravitational/teleport.e/issues/6759
Changelog: Add paginated API ListSAMLConnectors, deprecate GetSAMLConnectors