Fix Discovery boot when Proxy is not yet available (#60457)
If teleport is started with just Auth and Discovery in the same host,
the Public Proxy Addr is not available.

Even tho this means that Discovery will not be able to enroll Servers,
this is a valid set up.

When the Proxy becomes available, the next Discovery iteration will
correctly enroll servers.