discovery/access-graph: Restructure EKS log watcher
Move the EKS log watcher into the resource watcher as having them
separate meant they could run on different auth servers, and when that
happens, the Go channel between them will not work.

This is arguably a better structure as the EKS log watcher was already
coupled to the resource watcher. This makes it clearer.