[v18] EC2 auto discovery: allow usage of pre-defined `AWS-RunShellScript` SSM Document (#60434)
* Use AWS-RunShellScript as the default SSM Doc in EC2 Auto Discover

* escape url values

* Fix Discovery boot when Proxy is not yet available (#60457)

If teleport is started with just Auth and Discovery in the same host,
the Public Proxy Addr is not available.

Even tho this means that Discovery will not be able to enroll Servers,
this is a valid set up.

When the Proxy becomes available, the next Discovery iteration will
correctly enroll servers.