Fix Var usage (#60523)
Prepare for the migration of the Vale linter for Var component usage to
`remark` by editing a Var that only has a single occurrence. This linter
ensures that each Var has at least two occurrences on a page so we can
take advantage of the reused Var values or provide instructions on
assigning a Var.