discovery: Remove EKS audit logs from EKS fetcher
Remove checking EKS clusters for audit log fetching from the EKS fetcher
and move it up into the AWS resource sync. This has slightly more
overhead in that we need to reconstruct a map to check if it should have
audit logs fetched, but the code is much cleaner as a result.

Get rid of the clwClient interface as it embedded only a single other
interface - its a bit redundant even if it does follow the pattern of
the other AWS service clients.